refactor(emote add): Separating emote-related code
Moving to bot/cogs/emote.py and add a url arg to add emotes
